move version check to activity
